requirements.txt
anthropic>=0.40.0
openai
PyQt6
qasync
pynput
mss
Pillow
